Migrate generic hooks / add expiry field
