+++ src/config/sections/index.ts
@@ -1,2 +1,3 @@
 export * from "./cache";
 export * from "./queue";
+export * from "./generichooks";
